backend/backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.34.0,<0.35.0
python-dotenv>=1.1.0,<2.0.0
openai>=1.82.0,<2.0.0
python-multipart>=0.0.9,<0.1.0
requests>=2.32.0,<3.0.0
pydantic>=2.11.0,<3.0.0
pydantic-core>=2.23.0
Pillow>=11.3.0
pytesseract>=0.3.10,<0.4.0
numpy>=2.2.0,<3.0.0
scipy>=1.16.1,<2.0.0